requirements.txt
Brotli @ file:///Users/ec2-user/ci_py311/brotli-split_1678379922007/work
click @ file:///Users/ec2-user/ci_py311/click_1678318124837/work
contourpy==1.1.0
cycler==0.11.0
dash @ file:///Users/ec2-user/ci_py311/dash_1678410845018/work
dill==0.3.6
diskcache==5.6.1
Flask @ file:///Users/ec2-user/ci_py311/flask_1678380788105/work
Flask-Compress @ file:///Users/ec2-user/ci_py311/flask-compress_1678391154931/work
fonttools==4.40.0
intel-openmp==2023.1.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///Users/ec2-user/ci_py311/jinja2_1678317138271/work
kiwisolver==1.4.4
MarkupSafe @ file:///Users/ec2-user/ci_py311/markupsafe_1678316973889/work
matplotlib==3.7.1
mkl==2023.1.0
multiprocess==0.70.14
numpy==1.25.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly @ file:///Users/ec2-user/ci_py311/plotly_1678382250051/work
psutil==5.9.5
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
scipy==1.11.1
six==1.16.0
tbb==2021.9.0
tenacity @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cct5tq4h70/croot/tenacity_1682972283326/work
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_03acd6t6ca/croot/tqdm_1679561866522/work
tzdata==2023.3
Werkzeug @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_fegt3z60s3/croot/werkzeug_1679489762631/work
